+++ CHANGELOG.md
@@ -5,8 +5,10 @@ See the [keepachangelog.com description](https://keepachangelog.com/en/1.0.0/).
 ## 0.12.0 - 2026-02-21
 
 * Added
+  * PnP column selector includes the footprint
 * Changed
   * HTML report diff colors changed to darker
+  * HTML report diff takes the PnP footprint+comment against the BOM comment
 * Deprecated
 * Removed
 * Fixed